Add Dockerfile, wrapper script to build from container, and documentation
Building an OS image from within a container is a bit more complicated
than usual (eg. building an application), as it requires additional
privileges. In here, we rely on debos to build an OS image, and debos
needs to be able to start a QEMU VM from within the container. Therefore
we need to pass some additional arguments to the 'podman/docker run'
command.

A few words on the 'podman/docker run' options:
* `--log-driver none`: otherwise stdout also ends up in the journal
  (only applies to podman).
* `--net host`: make it easier to use a caching proxy on the localhost.
* `--user $(stat ...)`: we use `stat` rather than `id -u` to guess the
  user id, so that even when the script is called with sudo, we get the
  user id right (`id -u` gives zero if the script is run with sudo,
  obviously).

Note that there's no need to be root from within the container (the
script `build.sh` even bails out in this case). However the container
engine needs to be running as root. This is more or less the default
with docker (rootless exists but it's a fairly recent addition), so no
need to care. However this is not necessarily the default with podman,
it depends on whether you run it as 'podman' or 'sudo podman'.

Running rootless podman doesn't really fly, it leads us into permission
issues that are due to the combination of running rootless and using the
options --user and --volume, it seems. This is described at:
<https://www.redhat.com/sysadmin/user-namespaces-selinux-rootless-containers>.

In any case, and from my experience, building an OS image from within a
rootless podman, when it works, is much slower than with a root podman.
So I don't see any reason to try too hard to support this scenario.

Credits for this commit go to:
- <https://gitlab.apertis.org/infrastructure/apertis-image-recipes> (in
  particular, check the "Building in Docker" section of the README).
- <https://gitlab.com/ubports/core/rootfs-builder-debos> (in particular,
  check the `debos-*` scripts).

Fixes: #7