Use GB rather than GiB for size
It's not exactly a nitpick. debos uses github.com/docker/go-units to
parse the size, and docker/go-units is fundamentally broken, as it
parses both GiB anb GB as GB. It's a known issue that's been there
forever, and will probably never be fixed because backward compat
(that's my guess).

Cf. <https://github.com/docker/go-units/issues/31>, or better
<https://github.com/docker/go-units/pull/27#discussion_r158336500>.

So let's make sure that we talk about GB, because that's what happens in
practice.