Fix generate-* scripts to support a dash in version
In previous commit b912fd61, we started to set the default version
according to the branch, meaning that the version might now be
'last-snapshot', meaning that the version might contain a dash.

The code that parses the file name and extract the version didn't expect
that. This is now fixed. Note that there's also a dash in the version
for the weekly images, so it's REALLY something that needs to be
supported.

The other consequence, though, is that now we don't expect a dash in the
image type. It's OK for Kali virtual images for now, as the only image
types that exist are 'vmware' or 'virtualbox'.