Redefine the generic image to be in the OVF format
With this commit, the "generic" image is now defined as:
- having all the virtualization packages installed
- exported to OVA (ie. VMDK disk, OVF metadata file and manifest)

We now use the "MonolithicSparse" VMDK subformat (the default when
converting to vmdk format with 'qemu-img convert'). No more gzip
compression, because gzip doesn't compress enough, and once something is
compressed with gzip, it can't be compressed further with something else
(like xz).

Therefore it's a better approach to leave the disk uncompressed, and the
compress the OVA archive with something else, like 7zip.

You might notice that there's a 'raw' image in the recipes, but it's not
yet exposed in the build script. This is on the todo list.

This commit also updates the image descriptions in usage and readme.