Generate disk uuid when nothing is set
It seemed like a good idea to get the disk uuid correctly, as it was
introduced in 0862a24c.

But I guess that at the time, I tested on a vmdk disk that must have
been created by exporting a VirtualBox VM to ova. Because right now, I
don't see any trace of a disk uuid in the vmdk created by qemu-img, and
also looking at the code of qemu-img, there's no mention of a disk uuid
for the vmdk disk format.

However we need to set something, if the disk uuid is left empty,
VirtualBox fails to import the VM.