Add .tar.gz extension to ROOTFS variable
Not consistent with IMAGE (which doesn't include the extension). But
it's simpler to include the extension to ROOTFS, and it's simpler to NOT
include the extension to IMAGE. So, be it.

Since gz compression is the default for the pack action, there's no need
to write it down in the recipe.