Add VERSION to the rootfs name
Add VERSION, for clarity, but also to fix stuff.

Before this commit, when the rootfs name was simply rootfs-$ARCH.tar.gz,
and when building an image based on this rootfs, the build script had no
way to know whether the rootfs was built from Kali rolling or Kali
last-snapshot, and it just used whatever value was set for VERSION
(either the default value, or what was set by user on the command-line).

In practice, it means that you might build a `last-snapshot` rootfs with:

    ./build.sh -b kali-last-snapshot -t rootfs
    # -> creates rootfs-amd64.tar.gz

And then build an image based on that with:

    ./build.sh
    # -> creates an image named...

And the image would be named `kali-linux-rolling-generic-amd64.img`.
Wrong wrong wrong!

Therefore we must also include the VERSION in the rootfs name, so that
it's set properly in the image built from the rootfs.