Make rootfs reuse explicit via -r
Initially, I wanted the build.sh script to be "smart", and I thought it
could detect existing rootfs, and propose to reuse it, so that basically
the workflow could be:

  1. Build a rootfs with `./build.sh -t rootfs`
  2. Just run './build.sh -t XYZ' to build an image based on the rootfs

As it turns out, it's simple like that *only* when building with default
build options. However, if you build a rootfs for eg. arch=arm64 and
branch=kali-last-snapshot, then to reuse this rootfs you'll always need
to pass `-a arm64 -b kali-last-snapshot` so that the build system can
find the rootfs.

I think it will be error-prone, because if you build a rootfs today, and
come back tomorrow to build images based on that, all you might remember
is that the build.sh script is "smart" and automatically detects and
reuses existing rootfs, hence you just type `./build.sh` and press
Enter.  One hour later, you find out that you built a amd64 rolling
image, and your arm64 last-snapshot rootfs was not used at all. It might
not be immediately obvious that you need to add `-a arm64 -b
kali-last-snapshot` for the script to find the rootfs.

So, it's just my feeling, but I think it won't be very convenient to
use. So good-bye smartness, hello dumb code. From now on, if you want to
reuse a rootfs to build an image, you need to pass '-r ROOTFS'. Easy to
type thanks to completion, easy to remember (I hope).

It's still not perfect, actually it's even quite limited, as the code
needs to parse the filename to get the right values of VERSION and
ARCH. It means that you're not allowed to rename the rootfs, you must
keep it as it was named by the build script.

It should be easy to improve that though, so if need be we'll rework and
make it more flexible.