Prevent invalid usages when reusing rootfs
When reusing a rootfs, ARCH and VERSION are taken from the rootfs
filename, hence passing -a or -v at the same time makes no sense.

Let's check that and prevent invalid usages. It means that the code must
be a bit more sophisticated, we need to know whether the variable ARCH
was set from the command-line or not.