generate-ovf: Set ovf:version
VMware isn't super happy if we provide only `ovf:id="96"`, although it's
just a warning that doesn't appear in the GUI. Still, it doesn't accept
it and instead set the ID to "Other (32-bit)" or "Other Linux (64-bit)"
(not sure).

We can do better than that and also provide the field `ovf:version`.
However, it's still tricky. The version needs to be know for the
hardware family specified in the OVF (ie. the vmx-X).

For example, Debian 10 wasn't known at the time of `vmx-10`, therefore
VMware will fail to recognize it during import, and it will still fall
back to "Other Linux (64-bit)". However with a more recent version of
the hardware family (eg. `vmx-19`), VMware is happy.

For reference, here's the kind of warnings I've seen:

~~~~
+ <Warning>
+ <Type>
+ vim.fault.OvfMappedOsId
+ </Type>
+ <LocalizedMsg>
+ The specified operating system identifier &apos;Debian_64&apos; (id: 96) is not supported on the selected host. It will be mapped to the following OS identifier: &apos;Other (32-bit)&apos;.
+ </LocalizedMsg>
+ <Arg type="int" fieldName="ovfId">
+ 96
+ </Arg>
+ <Arg type="string" fieldName="ovfDescription">
+ Debian_64
+ </Arg>
+ <Arg type="string" fieldName="targetDescription">
+ Other (32-bit)
+ </Arg>
+ </Warning>
~~~~