generate-ovf: Make sound card optional
VMware can't import it. Apparently there's no sound card in the spec,
and the ResourceType 35 is specific to VirtualBox, it seems.

A OVF created by VirtualBox looks like that:

      <Item>
        <rasd:AutomaticAllocation>false</rasd:AutomaticAllocation>
        <rasd:ElementName>sound</rasd:ElementName>
        <rasd:InstanceID>7</rasd:InstanceID>
        <rasd:ResourceSubType>ensoniq1371</rasd:ResourceSubType>
        <rasd:ResourceType>35</rasd:ResourceType>
        <rasd:AddressOnParent>3</rasd:AddressOnParent>
        <rasd:Caption>sound</rasd:Caption>
        <rasd:Description>Sound Card</rasd:Description>
      </Item>

While a OVF created by VMware looks like that:

      <Item ovf:required="false">
        <rasd:AutomaticAllocation>false</rasd:AutomaticAllocation>
        <rasd:ElementName>sound</rasd:ElementName>
        <rasd:InstanceID>6</rasd:InstanceID>
        <rasd:ResourceSubType>vmware.soundcard.ensoniq1371</rasd:ResourceSubType>
        <rasd:ResourceType>1</rasd:ResourceType>
      </Item>

The funny thing is that VMware will fail to import the sound card even
for the second snippet, the one generated by VMware. Unless I missed
something...

So it seems that the easiest is just to set required=false, to ensure
that VMware will not emit a warning while importing the OVA.

For reference, the warning during import is:

~~~~
+ <Warning>
+ <Type>
+ vim.fault.OvfUnsupportedType
+ </Type>
+ <LocalizedMsg>
+ Line 124: OVF hardware element &apos;ResourceType&apos; with instance ID &apos;7&apos;: No support for the virtual hardware device type &apos;35&apos;.
+ </LocalizedMsg>
+ <Arg type="int" fieldName="lineNumber">
+ 124
+ </Arg>
+ <Arg type="string" fieldName="name">
+ ResourceType
+ </Arg>
+ <Arg type="string" fieldName="instanceId">
+ 7
+ </Arg>
+ <Arg type="int" fieldName="deviceType">
+ 35
+ </Arg>
+ </Warning>
~~~~