Disable OVA creation for generic image
Cf. previous commits: now the generic image is an uncompressed VMDK
disk, that we'll compress with 7zip for best compression ratio. At this
point it doesn't make much sense to put the various files in a tar
archive, and then compress it with 7zip. It's more straightforward to
simply put all the files directly in a 7zip archive.

Upon decompression, user will obtain 3 files: .ovf, .mf and .vmdk, and
then they can import the appliance by importing the .ovf. This mode of
distribution (set of files rather than ova archive) is allowed by the
OVF spec. In fact, when choosing to import a VM, both the extensions
.ova and .ovf are allowed in eg. VirtualBox or VMware.

Another reason for this change is that VMware doesn't support importing
an OVA with an uncompressed VMDK disk. Apparently when importing an OVA,
it only supports the "streamOptimized" VMDK format.

However there's no such limitation if we have a set of files and import
by selecting the .ovf: in this case VMware is happy with an uncompressed
VMDK.