Make default version dynamic
We can't set the default version straight at the beginning of the
script, because the default value is derived from the branch value, and
the branch value might be set by user. In such case, the default value
for version, as defined at the beginning of the script, is wrong.

So use a function rather than a variable, to be sure that the default
value is right. Also do it for arch, even though it's not needed, but
for consistency, let's use functions (rather than variables) for default
values.

This commit fixes abdb16ac.