Don't set BRANCH, DESKTOP and MIRROR when building from a rootfs
Those variable don't make sense when building from a rootfs, since
BRANCH, DESKTOP and MIRROR have already been set in order to build the
rootfs. And indeed, those variables are simply not used by the script,
when building from an existing rootfs.

So let's be a bit more strict: let's set those variables only when it
makes sense, and fail if user sets it while building an image from an
existing rootfs. Hopefully it prevent some confusion.

This is also in preparation for the commit to come.