Decouple variant and format, replace generic with generic-{ovf,raw}
Up to now, user could just pass a "type" to decide what type of image
they wanted to build. This "type" will then decide two things:
- what goes into the image, ie:
  - what virt support packages to install
  - what extra post-install step are needed (if any)
- what format will the image be exported to

It's mostly fine, but for the generic image we want a bit more control.
In particular, there are two formats that make sense for the generic
image:

- "raw": ie. no export, just let the image be a raw image, don't export
  it to any fancy format, don't generate a metadata file. That's perfect
  if you are *me*, and you just want to build an image as fast as
  possible, and then run it with QEMU straight after.
- "ovf": ie. export to the OVF format, with a sparse VMDK disk, so that
  it's easy for virtualbox/vmware users to consume it.

At the moment, the generic image is exported to OVF only, and there's no
option to change that, so it's not really what I want.

We could introduce a new image type named "raw", however that's a bit
awkward, because the image type goes into the image filename, so you'd
have an image named "kali-linux-rolling-raw-amd64.img". But "raw"
doesn't tell you what kind of image it is, in this case it would be a
generic image with all virt packages installed.

We could also decouple the image type in two: the VARIANT and the
FORMAT. It gives more flexibility, but it also allows a lot of nonsense,
for example you could build a vmware variant, and then export it to the
virtualbox format. Doesn't make sense.

So this commit goes kind of halfway: we DO decouple the TYPE into
VARIANT and FORMAT, but we don't really expose those two options
separately, instead we keep it hidden as an implentation detail within
the build.sh script.

In practice, what happens now is that:
* user can keep providing a single keyword, eg.  type=virtualbox, and it
  translates to variant=virtualbox and format=virtualbox.
* user can also provide a tuple (variant, format), for example
  type=generic-ovf.

The README and usage message suggest the types that make sense, but they
don't explain all the combinations (variant, format) that are possible.
So we only expose what makes sense, but as developers we know that we
can do a bit more.

Bonus points!

The script export-generic was split in two: export-ova.sh and
export-ovf.sh.

* export-ovf.sh is: vmdk sparse disk, ovf metadata, mf file. Optionally
  zipped with 7disk.
* export-ova.sh is: vmdk gzipped disk, ovf metadata, mf file. Packed in
  a tar archive. NOT zipped with 7disk in any case, because the disk is
  already compressed, and re-compressing doesn't save any space.