Finish the work on the VMware image
Some changes worth noting below.

Various uuids get generated if they're not in the .vmx file, so we don't
need to bother. I leave the functions to generate those for reference,
but they are not used anyway.

The 'extended config' aka .vmxf gets generated automatically, the NVRAM
aka .nvram as well. We still give a name for the nvram file, but it's
not mandatory, by default it would be simply called 'nvram' and that's
fine.