Use uppercase for desktop and packages: -D and -P
A rough attempt to split options in two (build options and customization
options), so that we can use lowercase and uppercase letters, as getopts
doesn't support long options and we're running out of letters.

The distinction is between the two categories is not really clear cut,
but we'll live with that.