Work on scratch size, default to 45G by default
Check how much scratch size is needed to build an image with the
different toolsets (default, large and everything).

Default to something big enough for the everything toolset + some extra
space, just in case someone builds an everything image + extra packages.

Under the hood, the scratch area is used to create the disk in the VM,
and the disk itself is a sparse file on the host, so the space is
allocated only when it's needed. So it doesn't matter if we default to
something way too big for the normal use case.