Add headless desktop environment (work in progress)
This commit works out which metapackage to install for the headless
image. Sure, the metapacakge kali-linux-headless is installed. But
there's a bit more to it.

First, for the headless image, we don't install any toolset by default.
This is in contrast to all the other images that come with a desktop
environment, and where we install kali-linux-default and
kali-tools-top10 by default. This being said, nothing prevents the user
to pass -S <whatever> to install a toolset, but then I suspect that it
will pull in the whole graphics stack, not sure if we can call that
headless at this point...

Second, for the headless image, when we install virt support, we must be
careful not to pull in the whole graphics stack. For now, I decided to
check if a kali-desktop-* package is installed. If yes, we install virt
support as usual (ie. with dependencies on X11). If no, we install the
virt support that doesn't depend on X11.

At this point I'm not even sure this is really the right approach.

Anyway. Known issue: for now the network is not enabled, so you won't
have any network connectivity out of the box with this image. I need to
put a bit more work in this.