Fill unallocated blocks with zeroes (only if -z option was set)
It's a bit tricky, as debos doesn't give us a scenario where we can
access all the partitions of the disk while they are unmounted. And
zerofree can't work on a partition that is already mounted.

Option 1: We can unmount the partition by ourself. But then debos
complains at the end of the action, when it tries to unmount the
partition and finds out that it's not mounted. So we need to mount it
again afterward, just so that debos can unmount it.

Option 2: Turns out that zerofree can work if the partition is mounted
read-only, so we can make it a bit shorter and simply remount the
partition read-only if ever it's mounted. It's enough for zerofree to
work, and it keeps debos happy as it can unmount it at the end of the
action.

Aside from that, it's worth noting that we do the zerofree only if the
`-z` option is given, ie. only when user wants to zip the image. So why?
The point of zerofree is to produce smaller images, when converting with
qemu-img to another format, or when we just plan to zip the raw image.
This is useful if we plan to distribute the image.  However if we're
building the image for local use, what we likely want is that the build
takes as little time as possible, and we don't care about the little
benefit from zerofree. So we rely on the `-z` option to "guess" if the
images are meant to be distributed or not.