Dig holes in the raw image before creating the bmap file
"Digging holes" means that sequences of zeroes that are long enough are
replaced by a hole in the sparse file. It makes the image file smaller
on disk. It doesn't make the size of the zippped image smaller though,
as zeroes are compressed away anyway. But then the image will be smaller
after decompression (compared to without digging holes), since bmap
tools recreate the holes.

In other words: for users, it doesn't change the download size, but it
reduces the image size on disk after decompression, assuming that they
decompress the image with bmap-tools (which is a must for a raw image).
So it's worth doing it.

"Digging holes" in the sparse file works best if zerofree was run on the
filesystem beforehand. It is the case, but it's not obvious, since
zerofree is run from the recipe before exporting.

Note that it seems that UBports had issues with this, cf the 2 commits:
- https://gitlab.com/ubports/development/core/rootfs-builder-debos/-/commit/3456b215
- https://gitlab.com/ubports/development/core/rootfs-builder-debos/-/commit/c5f572b8