Rework options for variant and format
This is a second take on commit 3fb5c759.

The idea to have a TYPE that bundles the couple VARIANT-FORMAT turned
out to be a bad idea, in the sense that the VARIANT is part of the name
of the image, but for the caller, the value of VARIANT is not known,
unless one looks into the script to see how the TYPE is unpacked into a
VARIANT and a FORMAT.

It means that the name of the image is not predictable, and for scripts,
that's a problem.

So we now expose both VARIANT and FORMAT, but with a twist: the default
format is chosen according to that VARIANT, so that user mostly
shouldn't have to care about it and shouldn't have to use this option.

So we keep the simplicity, while being more explicit, which is need for
automatic runs of the script.