Rewrite build-in-container.sh
Turns out it's OK to run rootless podman, it doesn't seem to cause any
performance hit, so let's enable it. The only gotcha is that:
- rootless podman: the user within the container must be root
- root podman: the user within the container must be same as non-root
  user on the host.

Also, with this commit we log the podman/docker commands that are run,
it will be easier to help others if things don't work on their side.