Save timezone settings in the debconf database
At the moment, here's what we have in the debconf database:

  # debconf-show tzdata | grep '^*'
  * tzdata/Areas: Etc
  * tzdata/Zones/Etc: UTC

However this is not correct, as the defaults are US/Eastern.

With this commit, after configuring the system, we save the values in
the debconf database. And here's how it looks like as a result:

  # debconf-show tzdata | grep '^*'
  * tzdata/Areas: America
  * tzdata/Zones/America: New_York
  * tzdata/Zones/Etc: UTC

Note that America/New_York is equivalent to US/Eastern, as we can see if
we resolve the symlink /usr/share/zoneinfo/US/Eastern.  Interestingly,
calling dpkg-reconfigure doesn't only update the database, it seems
that it also updates the files in /etc, as they now point to
America/New_York instead of US/Eastern:

  # cat /etc/timezone
  America/New_York
  # readlink /etc/localtime
  /usr/share/zoneinfo/America/New_York

Also, note that I tried another approach: to simply use the existing
script scripts/preseed.sh to pre-configure the system before we install
tzdata.  However tzdata is already installed when the preseed.sh script
runs (it's installed by debootstrap), so answering the debconf questions
at this point is too late. Then I tried running `dpkg-reconfigure
tzdata` to "apply" the new values that I just updated in the debconf
database, but what happens is exactly the opposite: instead
dpkg-reconfigure picks the values from the files in /etc, and update the
values in the database based on that. Not sure if it's a bug or the
expected behavior, cf. this comment that doesn't really answer:
<https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=352697#27>