Add first-boot service for grub-pc and debconf
We need to set grub-pc/install_devices in the debconf database,
otherwise the first time the package will get updated, dpkg will ask
where to install GRUB. This is confusing for users.

It's a bit tricky though, because the disk where GRUB is installed is
/dev/vda during the build, but then when the VM is run, it's probably
something different. It's /dev/sda when running with QEMU, and maybe
it's something different in VirtualBox or VMware. So we can't really
set this value at build time.

Note that if we were booting via UEFI, this wouldn't be needed, as GRUB
would be installed in the EFI partition, as a regular file, and there
would be no need to fiddle with block devices.

So the approach taken here is to add a script that will set the value
at first boot. The script is started from a systemd service and it's
meant to be executed only once. We remove the script after use, in order
to leave a clean /usr/local/sbin directory.