Install core and desktop metapackages in the same transaction
This is how it's done when installing Kali via the installer, or when we
create a Kali live image. The difference between installing both
metapackages in the same APT transaction, or in two different
transactions, is that APT might come up with slightly different lists of
package to install (due to conditional dependencies).

I don't want to introduce this kind of subtle differences in the VM
images, so, just to be on the safe side, let's keep installing both
metapackages in the same APT transaction.

This commit reworks the change introduced in
c0eaf1725180490ec67921509b940d7c1fc588b8.