Use '-D none' (instead of headless) to build an image with no desktop
The term headless actually refers to a metapackage that provides tools
that are useful for a Kali install without desktop. But this is not a
"desktop metapackage", instead is a "toolset metapackage".

With this commit, let's clear up the confusion. Let's use "none" as the
keyword that means "no desktop". At the same time, add "headless" to the
list of toolsets, and make it the default when there's no desktop.

To sum up:

    ./build.sh -D <something>
    -> gives you Kali with a desktop, and kali-linux-default tools

    ./build.sh -D none
    -> gives you Kali with no desktop, and kali-linux-headless tools

As one can see, the default is always to install tools, because we want
to build a useful Kali image by default. If you don't want tools, you
need to pass "-S none" explicitly.