Don't install kali-tools-top10 explicitly
In the last commit, we added "headless" as a toolset. If users builds an
image with the kali-linux-headless toolset, it's likely that they didn't
install a desktop, and in such case, we don't want to install the top10
metapackage.

Actually there's no need to mention the top10 metapackage explicitly, as
it's pulled in by kali-linux-{default,large,everything}, so it will be
installed anyway.