Breaking change in build.sh!  --  -T becomes -Z, -S becomes -T
Setting the default toolset is arguably one of the most useful option
(at least, I use it all the time), so let's give it the most intuitive
letter (ie. -T for 'toolset' or 'tools').

Then for timezone, fall back to the second most intuitive letter, since
-T is taken for toolset: let's use -Z then.

Sorry for the disruption