Install usr-is-merged to fix build failure
Somehow apt chokes while trying to resolve the set of standard packages.
This seems to be due to the following dependency chain:

  dbus -> init-system-helpers -> usrmerge | usr-is-merged.

I couldn't understand why apt refuses to install usrmerge (installing it
manually works, but it requires listing all the dependencies that it
complains about on the command-line). I really couldn't understand what
was wrong with those deps.

In any case, it turns out that usr-is-merged *should have been
installed*, that's what debootstrap is supposed to do. However, this has
been disabled in a recent debos commit, cf:
* https://github.com/go-debos/debos/issues/361
* https://github.com/go-debos/debos/pull/362/commits/b3c1f76b

So let's install usr-is-merged manually (as the debos workaround is not
for our use-case anyway), that brings us back in line with a standard
debootstrap chroot, which is surely a good thing.

From this point the problem disappears.

Fixes: #23