Revert "Retry when fallocate fails"
This reverts commit a6ab3c8c33edffd085e4873ebf8b4e53264b2c2f.

The fallocate failures, that I used to see on a regular basis, seem to
be caused by lack of RAM in the VM.  It's a bit weird, since this script
runs just after filesystem-deploy, and it shouldn't need any RAM at all
(it allocates a file on a virtual hard disk, and the way we run debos,
the scratch area is backed by the disk, so why need RAM?).

I wonder if the big 'cp -a' that is done by filesystem-deploy actually
fills the RAM, and it takes a while for this data to be flushed to disk,
and so when we run a script just after, somehow there's no RAM avail...

Maybe I could validate that by just starting the script with a sleep.
But for now, my solution was to incread the VM RAM to 8G (instead of the
default 4G that is hardcoded in build.sh), and that seems to do the
trick.

So let's drop all thos debug messages, because they're useless anyway,
and move on.