Use that --artifactdir option from debos
As a result, running the various export-* scripts by hand (which might
be useful to develop and debug) will no longer work out of the box.
Instead, one must set environment variables before.

I suppose that the right way to call those scripts now will be:

    RECIPEDIR=$(pwd) ARTIFACTDIR=$(pwd)/images \
        ./scripts/export-<whatever>.sh ...

So what's the benefit, compared to passing the artifact dir as part of
the image name?

First it's a bit cleaner, to have both artifactdir and filename in
separate variables, rather than juggling with dirname and basename.

It also allows caller to use the --artifactdir option of debos. Before
this commit, I supposed it worked, but it would always create a subdir
"images" within artifactdir, while now, if you pass --artifactdir, the
output dir is really the artifact dir. So the OUTDIR in build.sh is now
configurable via --artifactdir option of debos, which is a nice thing
(we allow caller to pass option to debos, so it's nice if we really
support it).

Other than that, I'm not sure there's such a clear benefit. Time might
tell if it was a good move or not.