Create a $ARTIFACTDIR/.artifacts file after a successful build
This file lists the build artifacts produced by the build, to some
extent. The main benefit is a better output for user after the build
succeeds. It also makes it easier for scripts that call build.sh, as
they can discover the names of the build artifacts.

I say "to some extent" because, first, we only match filenames in the
artifact dir, and not deeper. Most of the times, a build produces a set
of files, so that's fine, but in case you'd build a vmware image and
leave it uncompressed, the output will be a directory. And so, no, we
don't list what's in this directory, we'll just list the directory
itself.

Also, the implementation is a bit sloppy, on purpose, to keep it simple.
So we don't do a thorough bookkeeping of the build artifacts, instead we
just list the files named "$image.*" in the artifact dir after the build
finishes, check their timestamp, and keep only the files modified after
the start time of the script.

Meaning that the raw image itself, which was produced before running the
export-* scripts, won't be included (which can be a feature or a bug,
depending on what you expect). It also means that, if ever people run
two builds at the same time (eg. in parallel from two terminals), there
might be surprises. But do people really do that?